refactor: wrap out the participants name
