+++ src/main/java/domain/LadderHeight.java
@@ -1,0 +1,4 @@
+package domain;
+
+public class LadderHeight {
+}
