+++ src/main/java/domain/LadderWidth.java
@@ -1,0 +1,4 @@
+package domain;
+
+public class LadderWidth {
+}
